feat(model): Add entities

+++ src/main/java/madstodolist/repository/UsuarioRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
+public interface UsuarioRepository extends CrudRepository<Usuario, Integer> {
     Optional<Usuario> findByEmail(String s);
 }
